+++ packages/bpk-docs/src/layouts/DocsLayout/DocsLayout.js
@@ -29,6 +29,7 @@ const links = [
       { id: 'FORMS', route: routes.FORMS, children: 'Forms' },
       { id: 'CARDS', route: routes.CARDS, children: 'Cards' },
       { id: 'BADGE', route: routes.BADGE, children: 'Badge' },
+      { id: 'PANELS', route: routes.PANELS, children: 'Panels' },
       { id: 'ILLUSTRATIONS', route: null, children: 'Illustrations' },
       { id: 'IMAGES', route: null, children: 'Image' },
     ],